+++ pkg/service/service.go
@@ -17,8 +17,8 @@ func NewService(repo *repository.Repository) *Service {
 
 type User interface {
 	Balance(userId int) (*avitoTech.User, error)
-	TopUp(userId int, amount int) (*avitoTech.User, error)
-	Debit(userId int, amount int) (*avitoTech.User, error)
-	Transfer(userId int, toId int, amount int) (*avitoTech.User, error)
+	TopUp(userId int, amount float64) (*avitoTech.User, error)
+	Debit(userId int, amount float64) (*avitoTech.User, error)
+	Transfer(userId int, toId int, amount float64) (*avitoTech.User, error)
 	ConvertBalance(ans *avitoTech.User, currency string) (*avitoTech.User, error)
 }
