+++ user.go
@@ -3,5 +3,5 @@ package avitoTech
 type User struct {
 	Id      int `json:"-" db:"id"`
 	UserId  int `json:"user_id" db:"user_id"`
-	Balance int `json:"balance" db:"balance"`
+	Balance float64 `json:"balance" db:"balance"`
 }
